AIAI-128 re-org test: split demo tests and feature-centric tests
