+++ smart-orchestrator-test/src_test/com/axonivy/utils/smart/orchestrator/agent/message/system/TestAgenticProcessCallElementSystemMessage.java
@@ -1,4 +1,4 @@
-package com.axonivy.utils.ai.tools.test.output;
+package com.axonivy.utils.smart.orchestrator.agent.message.system;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,21 +12,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import com.axonivy.utils.ai.mock.MockOpenAI;
+import com.axonivy.utils.smart.orchestrator.client.OpenAiTestClient;
 import com.axonivy.utils.smart.orchestrator.connector.OpenAiServiceConnector.OpenAiConf;
 import com.axonivy.utils.smart.orchestrator.test.TestToolUserData;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 
 import ch.ivyteam.ivy.bpm.engine.client.BpmClient;
 import ch.ivyteam.ivy.bpm.engine.client.element.BpmProcess;
 import ch.ivyteam.ivy.bpm.exec.client.IvyProcessTest;
 import ch.ivyteam.ivy.environment.AppFixture;
-import ch.ivyteam.test.client.OpenAiTestClient;
 import ch.ivyteam.test.log.LoggerAccess;
 import dev.langchain4j.http.client.log.LoggingHttpClient;
 
 @IvyProcessTest(enableWebServer = true)
-class TestAgenticProcessCallElementOutput {
+class TestAgenticProcessCallElementSystemMessage {
 
   private static final BpmProcess AGENT_TOOLS = BpmProcess.name("TestToolUser");
 
@@ -35,24 +34,17 @@ class TestAgenticProcessCallElementOutput {
 
   @BeforeEach
   void setup(AppFixture fixture) {
-    fixture.var(OpenAiConf.BASE_URL, OpenAiTestClient.localMockApiUrl("output"));
+    fixture.var(OpenAiConf.BASE_URL, OpenAiTestClient.localMockApiUrl("systemMessage"));
     fixture.var(OpenAiConf.API_KEY, "");
-    MockOpenAI.defineChat(this::structure);
+    MockOpenAI.defineChat(this::repeat);
   }
 
-  private Response structure(JsonNode request) {
-    var messages = (ArrayNode) request.get("messages");
-    if (messages.size() == 1) { // tool response
-      return sendResource("response1.json");
-    }
-    if (messages.size() == 3) { // final response
-      return sendResource("response2.json");
-    }
-    return Response.serverError().build();
+  private Response repeat(@SuppressWarnings("unused") JsonNode request) {
+    return sendResource("response.json");
   }
 
   private static Response sendResource(String resource) {
-    try (InputStream is = TestAgenticProcessCallElementOutput.class.getResourceAsStream(resource)) {
+    try (InputStream is = TestAgenticProcessCallElementSystemMessage.class.getResourceAsStream(resource)) {
       return Response.ok()
           .entity(new String(is.readAllBytes()))
           .build();
@@ -62,13 +54,11 @@ class TestAgenticProcessCallElementOutput {
   }
 
   @Test
-  void structuredOutput(BpmClient client) {
-    var res = client.start().process(AGENT_TOOLS.elementName("structuredOutput")).execute();
+  void systemMessage(BpmClient client) {
+    var res = client.start().process(AGENT_TOOLS.elementName("systemMessage")).execute();
     TestToolUserData data = res.data().last();
-    assertThat(data.getPerson().getFirstName())
-        .isEqualTo("James");
-    assertThat(data.getPerson().getLastName())
-        .isEqualTo("Bond");
+    assertThat(data.getResult())
+        .contains("it's so hot, right?");
   }
 
 }
