+++ smart-orchestrator-test/src_test/com/axonivy/utils/smart/orchestrator/agent/message/system/response.json
