+++ smart-orchestrator-test/src_test/com/axonivy/utils/smart/orchestrator/agent/message/user/TestAgenticProcessCallElementQuery.java
@@ -1,10 +1,9 @@
-package com.axonivy.utils.ai.tools.test.filter;
+package com.axonivy.utils.smart.orchestrator.agent.message.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.ArrayList;
 
 import javax.ws.rs.core.Response;
 
@@ -13,21 +12,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import com.axonivy.utils.ai.mock.MockOpenAI;
+import com.axonivy.utils.smart.orchestrator.client.OpenAiTestClient;
 import com.axonivy.utils.smart.orchestrator.connector.OpenAiServiceConnector.OpenAiConf;
 import com.axonivy.utils.smart.orchestrator.test.TestToolUserData;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 
 import ch.ivyteam.ivy.bpm.engine.client.BpmClient;
 import ch.ivyteam.ivy.bpm.engine.client.element.BpmProcess;
 import ch.ivyteam.ivy.bpm.exec.client.IvyProcessTest;
 import ch.ivyteam.ivy.environment.AppFixture;
-import ch.ivyteam.test.client.OpenAiTestClient;
 import ch.ivyteam.test.log.LoggerAccess;
 import dev.langchain4j.http.client.log.LoggingHttpClient;
 
 @IvyProcessTest(enableWebServer = true)
-class TestAgenticProcessCallElement {
+class TestAgenticProcessCallElementQuery {
 
   private static final BpmProcess AGENT_TOOLS = BpmProcess.name("TestToolUser");
 
@@ -36,23 +34,21 @@ class TestAgenticProcessCallElement {
 
   @BeforeEach
   void setup(AppFixture fixture) {
-    fixture.var(OpenAiConf.BASE_URL, OpenAiTestClient.localMockApiUrl("tools.filter"));
+    fixture.var(OpenAiConf.BASE_URL, OpenAiTestClient.localMockApiUrl("query"));
     fixture.var(OpenAiConf.API_KEY, "");
-    MockOpenAI.defineChat(this::toolsFilter);
+    MockOpenAI.defineChat(this::query);
   }
 
-  private Response toolsFilter(JsonNode request) {
-    var tools = (ArrayNode) request.get("tools");
-    var toolNames = new ArrayList<String>();
-    tools.forEach(tool -> toolNames.add(tool.get("function").get("name").asText()));
-    if (toolNames.size() == 1 && "whoami".equals(toolNames.get(0))) {
-      return sendResource("response.json");
+  private Response query(JsonNode request) {
+    var userMessage = request.get("messages").get(0).get("content").asText();
+    if (userMessage.contains("ivy.session")) {
+      throw new IllegalStateException("given 'query' was not expaned, received: " + userMessage);
     }
-    return Response.serverError().build();
+    return sendResource("response.json");
   }
 
   private static Response sendResource(String resource) {
-    try (InputStream is = TestAgenticProcessCallElement.class.getResourceAsStream(resource)) {
+    try (InputStream is = TestAgenticProcessCallElementQuery.class.getResourceAsStream(resource)) {
       return Response.ok()
           .entity(new String(is.readAllBytes()))
           .build();
@@ -62,10 +58,10 @@ class TestAgenticProcessCallElement {
   }
 
   @Test
-  void toolFilter(BpmClient client) throws NoSuchFieldException {
-    var res = client.start().process(AGENT_TOOLS.elementName("reducedTools")).execute();
+  void expandQuery(BpmClient client) {
+    var res = client.start().process(AGENT_TOOLS.elementName("expandQuery")).execute();
     TestToolUserData data = res.data().last();
-    assertThat((String) data.get("result")).contains("whoami");
+    assertThat(data.getResult()).contains("Unknown User");
   }
 
 }
