+++ smart-orchestrator-test/src_test/com/axonivy/utils/smart/orchestrator/output/TestAgenticProcessCallElementOutput.java
@@ -1,4 +1,4 @@
-package com.axonivy.utils.ai.tools.test.query;
+package com.axonivy.utils.smart.orchestrator.output;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,20 +12,21 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import com.axonivy.utils.ai.mock.MockOpenAI;
+import com.axonivy.utils.smart.orchestrator.client.OpenAiTestClient;
 import com.axonivy.utils.smart.orchestrator.connector.OpenAiServiceConnector.OpenAiConf;
 import com.axonivy.utils.smart.orchestrator.test.TestToolUserData;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.ArrayNode;
 
 import ch.ivyteam.ivy.bpm.engine.client.BpmClient;
 import ch.ivyteam.ivy.bpm.engine.client.element.BpmProcess;
 import ch.ivyteam.ivy.bpm.exec.client.IvyProcessTest;
 import ch.ivyteam.ivy.environment.AppFixture;
-import ch.ivyteam.test.client.OpenAiTestClient;
 import ch.ivyteam.test.log.LoggerAccess;
 import dev.langchain4j.http.client.log.LoggingHttpClient;
 
 @IvyProcessTest(enableWebServer = true)
-class TestAgenticProcessCallElementQuery {
+class TestAgenticProcessCallElementOutput {
 
   private static final BpmProcess AGENT_TOOLS = BpmProcess.name("TestToolUser");
 
@@ -34,21 +35,24 @@ class TestAgenticProcessCallElementQuery {
 
   @BeforeEach
   void setup(AppFixture fixture) {
-    fixture.var(OpenAiConf.BASE_URL, OpenAiTestClient.localMockApiUrl("query"));
+    fixture.var(OpenAiConf.BASE_URL, OpenAiTestClient.localMockApiUrl("output"));
     fixture.var(OpenAiConf.API_KEY, "");
-    MockOpenAI.defineChat(this::query);
+    MockOpenAI.defineChat(this::structure);
   }
 
-  private Response query(JsonNode request) {
-    var userMessage = request.get("messages").get(0).get("content").asText();
-    if (userMessage.contains("ivy.session")) {
-      throw new IllegalStateException("given 'query' was not expaned, received: " + userMessage);
+  private Response structure(JsonNode request) {
+    var messages = (ArrayNode) request.get("messages");
+    if (messages.size() == 1) { // tool response
+      return sendResource("response1.json");
     }
-    return sendResource("response.json");
+    if (messages.size() == 3) { // final response
+      return sendResource("response2.json");
+    }
+    return Response.serverError().build();
   }
 
   private static Response sendResource(String resource) {
-    try (InputStream is = TestAgenticProcessCallElementQuery.class.getResourceAsStream(resource)) {
+    try (InputStream is = TestAgenticProcessCallElementOutput.class.getResourceAsStream(resource)) {
       return Response.ok()
           .entity(new String(is.readAllBytes()))
           .build();
@@ -58,10 +62,13 @@ class TestAgenticProcessCallElementQuery {
   }
 
   @Test
-  void expandQuery(BpmClient client) {
-    var res = client.start().process(AGENT_TOOLS.elementName("expandQuery")).execute();
+  void structuredOutput(BpmClient client) {
+    var res = client.start().process(AGENT_TOOLS.elementName("structuredOutput")).execute();
     TestToolUserData data = res.data().last();
-    assertThat(data.getResult()).contains("Unknown User");
+    assertThat(data.getPerson().getFirstName())
+        .isEqualTo("James");
+    assertThat(data.getPerson().getLastName())
+        .isEqualTo("Bond");
   }
 
 }
